changed the behaviour of checkgrad
checkgrad usd to check the passed string (for name matching) against the
list of _get_param_names(). Then it would index along
_get_param_names_transformed()!

this led to inconsistensies when fixed or tied variables were used,
which screwed up the ordering of the variable names.

We now match against _get_param_names_transformed().