fix: dev: cython import errors
This commit fixes issues observed in Windows where some
cython modules are successfully imported, and some are not.
This causes the global config cython.working to be inconsistent,
which causes import errors when unavailable cython modules
are tried to be imported (example
https://github.com/SheffieldML/GPy/issues/266). This commit uses
a separate flag for each module to fix the issue.